requirements.txt
langgraph
langchain-openai
python-dotenv
tavily-python
pandas
scikit-learn
seaborn
